hmx: elaborate expr
This commit makes `infer ~env e` return an elaborated expression.

Currently the source expression and elaborated expression share their
representation. We might change it later but for now the only
elaboration happens by annotating let-expressions with type schemes.

The tests changed to assert elaborated expressions instead of types
only. This allows to see the types of inner let-expressions which is
going to be useful when we extend hmx with new types of constraints.